refactor(service/database): 将 ProjectTask 中的 order 字段类型从 Long 改为 Integer
- 在 ProjectTaskDTO、ProjectTaskPO 和 ProjectTaskTableSchema 中均进行了类型修改
- 此修改统一了 order 字段的类型，提高了代码的一致性和可维护性

+++ database/src/main/java/com/advantest/demeter/database/schema/ProjectTaskTableSchema.java
@@ -28,7 +28,7 @@ public record ProjectTaskTableSchema(
         Long projectId,
 
         @Column(name = "`ORDER`")
-        Long order,
+        Integer order,
 
         @Column(name = "CREATOR_ID", nullable = false)
         Long creatorId,
